old-removed-requirements.txt
google-cloud-logging==1.4.0
google-cloud-error_reporting==0.29.0
oauth2client==4.1.2
six==1.11.0
requests[security]==2.18.4
